requirements.txt
requests==2.31.0
requests-file==2.0.0
requests-html==0.10.0
requests-toolbelt==1.0.0
python-dotenv==1.0.0